BE-747 Refactoring initialise sequence in FabricClient for fabric-v2.1
