+++ app/platform/fabric/service/NetworkService.js
@@ -36,7 +36,7 @@ class NetworkService {
 			logger.debug('Network list ', networkName);
 			networklist.push({
 				name: networkName,
-				authEnabled: clientObj.instance.fabricGateway.getEnableAuthentication()
+				authEnabled: clientObj.instance.getEnableAuthentication()
 			});
 		}
